requirements.txt
gTTS
TTS 
transformers
torch
scipy
Pillow
librosa
soundfile
numpy
gradio
num2words